[FEAT] 네이버 즐겨찾기 저장

+++ src/components/features/ExtractedPlacesList/ExtractedPlacesList.tsx
@@ -5,21 +5,25 @@ import { Chip } from '@/components/common/Chip';
 import { Icon } from '@/components/common/Icon';
 import { ListCard } from '@/components/common/ListCard';
 import { Body2, Body3, Body4 } from '@/components/common/Typography';
+import { findyIconNames } from '@/constants/findyIcons';
+import { useNaverBookmark } from '@/hooks/api/bookmarks/useNaverBookmark';
 import { useYoutubeBookmark } from '@/hooks/api/bookmarks/useYoutubeBookmark';
-import { YoutubeResponse } from '@/hooks/api/link/useYoutubePlace';
+import { ExtractResponse } from '@/hooks/api/link/useYoutubePlace';
 import { useAuth } from '@/hooks/auth/useAuth';
 import { useMarkers } from '@/hooks/common/useMarkers';
+import { Category } from '@/types/naver';
 
 import { Login } from '../LoginModal';
 
-type Props = { places: YoutubeResponse; onNext: () => void };
-export const ExtractedPlacesList = ({ places, onNext }: Props) => {
+type Props = { data: ExtractResponse; onNext: () => void };
+export const ExtractedPlacesList = ({ data, onNext }: Props) => {
   const [selectedIds, setSelectedIds] = useState<number[]>([]);
   const [isOpen, setIsOpen] = useState<boolean>(false);
 
   const { token } = useAuth();
   const { clearMarkers } = useMarkers();
-  const { mutate: bookmarkMutate } = useYoutubeBookmark(token);
+  const { mutate: youtubeMutate } = useYoutubeBookmark(token);
+  const { mutate: naverMutate } = useNaverBookmark(token);
 
   const handleToggleSelect = (id: number) => {
     setSelectedIds((prev) =>
@@ -32,53 +36,69 @@ export const ExtractedPlacesList = ({ places, onNext }: Props) => {
       setIsOpen(true);
       return;
     }
-
-    const savePlaces = {
-      ...places,
-      places: places.places
+    const savePlaces: ExtractResponse = {
+      ...data,
+      places: data.places
         .filter((place) => selectedIds.includes(place.id as number))
         // eslint-disable-next-line @typescript-eslint/no-unused-vars
         .map(({ id, ...placeData }) => placeData),
     };
 
-    bookmarkMutate(savePlaces, {
-      onSuccess: () => {
-        sessionStorage.clear();
-        clearMarkers();
-        onNext();
-      },
-    });
+    if (data.youtuberId) {
+      youtubeMutate(savePlaces, {
+        onSuccess: () => {
+          sessionStorage.clear();
+          clearMarkers();
+          onNext();
+        },
+      });
+    }
+    if (data.name) {
+      naverMutate(savePlaces, {
+        onSuccess: () => {
+          sessionStorage.clear();
+          clearMarkers();
+          onNext();
+        },
+      });
+    }
   };
   return (
     <div className="flex flex-col gap-4 p-3">
       <div className="flex flex-row gap-4 py-2">
-        {/* TODO : 이미지가 없을 경우 대체 이미지 추가 */}
-        <img
-          src={places.youtuberProfile}
-          className="w-12 h-12 rounded-full"
-          alt={`${places.youtuberName}의 프로필 이미지`}
-        />
-        <div className="flex flex-col ">
-          <Body2 weight="medium">{places.youtuberName}</Body2>
+        {data.youtuberProfile ? (
+          <img
+            src={data.youtuberProfile || ''}
+            className="w-12 h-12 rounded-full "
+            alt={`${data.youtuberName}프로필 이미지`}
+          />
+        ) : (
+          <Icon name={findyIconNames[0]} className="w-11 h-11" />
+        )}
+        <div className="flex flex-col">
+          <Body2 weight="medium">{data.youtuberName ?? data.name}</Body2>
           <div className="flex flex-row items-center gap-1">
-            <Icon name="location" size={20} />
-            <Body3 className=" text-gray-500">{places.places.length}</Body3>
+            <Icon name="location" size={17} />
+            <Body3 className="text-gray-500">{data.places.length}</Body3>
           </div>
         </div>
       </div>
       <ListCard>
         {/* TODO 컴포넌트화 */}
-        {places.places.map((item, index) => (
-          <>
+        {data?.places.map((item, index) => (
+          <div key={`${item.title}-${item.address}`}>
             <div
-              key={`${item.title}-${item.address}`}
-              className={`flex flex-row justify-between items-center ${index !== places.places.length - 1 && 'pb-2'}`}
+              className={`flex flex-row justify-between items-center ${index !== data.places.length - 1 && 'pb-2'}`}
             >
               <div className="flex flex-col gap-1 py-2">
                 <div className="flex flex-row gap-3 items-center">
                   <Body2 className="text-primary">{item.title}</Body2>
-                  {typeof item.category === 'object' && (
-                    <Chip variant="medium">{item.category.majorCategory}</Chip>
+                  {item.category && (
+                    <Chip variant="medium">
+                      {typeof item.category === 'object'
+                        ? (item.category as Category).majorCategory
+                        : item.category}
+                    </Chip>
                   )}
                 </div>
                 <Body4 className="pt-1" weight="normal">
@@ -92,8 +112,8 @@ export const ExtractedPlacesList = ({ places, onNext }: Props) => {
                 onClick={() => handleToggleSelect(item.id as number)}
               />
             </div>
-            {index < places.places.length - 1 && <hr className="border-dashed pt-2" />}
-          </>
+            {index < data.places.length - 1 && <hr className="border-dashed pt-2" />}
+          </div>
         ))}
       </ListCard>
       <Button variant="primary" size="large" onClick={handleSave}>
